[cffLib] Move desubroutinize code to new cffLib.transforms module
